Refactor: Phase 1 - Extract dispatcher utilities and basic search operations
This is the first phase of breaking down the monolithic dispatcher.ts file.

## Phase 1 Accomplishments

### ✅ Directory Structure Created
-  - Main dispatcher module directory
-  - Operation-specific handlers

### ✅ Utility Modules Extracted
- **logging.ts** - logToolRequest, logToolError utilities
- **validation.ts** - validateAttributes, validateResourceId functions
- **formatting.ts** - formatSuccessResponse utility

### ✅ Search Operations Extracted
- **operations/search.ts** - Basic search operation handlers:
  - handleBasicSearch
  - handleSearchByEmail
  - handleSearchByPhone
  - handleSmartSearch
  - handleSearchOperation

### ✅ Core Dispatcher Created
- **core.ts** - Main executeToolRequest function with modular design
- Handles: search, searchByEmail, searchByPhone, smartSearch, details, notes, createNote, getLists
- Clean separation of concerns with focused operation handlers

### ✅ Backward Compatibility Maintained
- **dispatcher.ts** - Updated to re-export from modular implementation
- All existing imports continue to work
- Public API unchanged

## Benefits Achieved

✅ **Reduced Complexity** - Core dispatcher reduced from 2,200+ lines to manageable modules
✅ **Better Organization** - Clear separation of logging, validation, formatting, and operations
✅ **Improved Maintainability** - Each module has single responsibility
✅ **Enhanced Testability** - Individual utilities can be unit tested in isolation
✅ **Zero Breaking Changes** - All existing functionality preserved

## Next Steps

- Phase 2: Extract remaining operation modules (companies, records, lists, notes, relationships, advanced-search)
- Add comprehensive unit tests for extracted modules
- Performance validation

## Issue Progress

Addresses Issue #238 - Phase 1 of dispatcher modularization complete.
